Addded Finalizer to MemoryMappedFile
Memory mapped files were not getting closed. TcImports attaches a dispose function to it's list of disposables to close the map, however, TcImports is never disposed in the "main" fsc.fs functions. So either TcImports needs to be disposed, a finalizer needs to be added to TcImports or a Finalizer needs to be added to MemoryMappedFile. Here the finalizer was added as close to the resources as possible.