Fix output location of referenced projects
Prior to this the referencing project file's path would be concatenated
to the relative path to the referenced project's output directory. If
either one of these was the same across the two projects (usually the
latter) this bug was masked.

Also added a test.

Also disabled PCL tests on Linux where PCL is normally not available and
these tests fail.