Move FCS project cracking out of process
This aims to fix many problems stemming from two main issues:

* FCS depends on MSBuild v12, and not all machines have this installed.
* It's hard to apply robust assembly redirects, as this is up to the
  application referencing FCS. These are needed to redirect MSBuild v4
  to whatever is being used (for custom MSBuild tasks), and also to
  allow redirecting to MSBuild v14 if necessary.

The API is preserved, with an additional function,
`GetProjectOptionsFromProjectFileLogged` to allow easier access to the
logs for presentation to users and bug-reporting.

The functionality is implemented by FCS calling a new console
application, FSharp.Compiler.Service.ProjectCracker.exe, and the
response is transmitted via serialization using JSON to types shared by
a common source file, so there is no binary dependency between the two.

The ProjectCracker static links to FSharp.Core, as it is in the same
directory as FCS.dll. Thus bundling a separate copy would force the
choice of version on all users of FCS, which is not acceptable.