Fix the look up of members added to a provided type by an intrinsic extension
This is part of the fix for https://github.com/Microsoft/visualfsharp/issues/49

The issue was that the functions in infos.fs responsible for the look up,
GetImmediateIntrinsicMethInfosOfType and GetImmediateIntrinsicPropInfosOfType,
perform a match on the type of metadata of the type. When the ProvidedTypeMetadata
case is executed, only that metadata is searched, ignoring any F# metadata for the
extension members.

This patch aims to fix that by returning the union of the ProvidedTypeMetadata and
F# metadata.