ensure Microsoft.FSharp.NetSdk.props makes it into VisualFSharpFull.vsix (#6942)
* ensure Microsoft.FSharp.NetSdk.props makes it into VisualFSharpFull.vsix

During a build/deploy in VS the target `BeforeBuild` wasn't getting called
on FSharp.Build.fsproj if up-to-date checks passed which meant the .props
file wasn't getting added to the @(None) item group.  The fix is to make
that target depend on `AssignTargetPaths` which is run in both the VS
build as well as the command line build.

* annotate ambiguous function